requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
python-dotenv>=1.0.0
Markdown>=3.5.0
Pillow>=10.0.0
